LeetCode/src/test/java/org/redquark/onlinejudges/leetcode/tree/PathSumTest.java
package org.redquark.onlinejudges.leetcode.tree;

import org.junit.jupiter.api.Test;
import org.redquark.onlinejudges.leetcode.common.TreeNode;

import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;

public class PathSumTest {

    private final PathSum testObject = new PathSum();

    @Test
    public void testHasPathSum() {
        TreeNode root = new TreeNode(5);
        root.left = new TreeNode(4);
        root.right = new TreeNode(8);
        root.left.left = new TreeNode(11);
        root.right.left = new TreeNode(13);
        root.right.right = new TreeNode(4);
        root.left.left.left = new TreeNode(7);
        root.left.left.right = new TreeNode(2);
        root.right.right.right = new TreeNode(1);
        int targetSum = 22;
        assertTrue(testObject.hasPathSum(root, targetSum));

        root = new TreeNode(1);
        root.left = new TreeNode(2);
        root.right = new TreeNode(3);
        targetSum = 5;
        assertFalse(testObject.hasPathSum(root, targetSum));

        targetSum = 0;
        assertFalse(testObject.hasPathSum(null, targetSum));
    }
}